MultiPropertyEditor: Avoid model manipulation on custom notifications
The model shouldn't be manipulated in custom notification event
handlers. Creating a view adds the view to the views list of the
model.

Instead of handling events by custom notifications, callback methods
are introduced.

MultiPropertyEditor doesn't take care of any custom notification
anymore, and they might be handled only in the main
PropertyEditorView.

Fixes: QDS-15715
Change-Id: I4744bb1dd46c40f0998b2aff748877d065738783
Reviewed-by: Marco Bubke <marco.bubke@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>