qmlls: allow the user to override the qmlls executable
Allow the user to override the qmlls executable. We use the --version
parameter of qmlls to know which command line parameters we can pass to
qmlls. Sadly, --version is only introduced in Qt 6.10 so users can't
override the qmlls executable with executables older than 6.10 (at the
time of writing this means that only the qmlls from the 6.10 beta and
the standalone version 0.2 can be used like this).

Fixes: QTCREATORBUG-32749
Change-Id: I6ccce60d59dc9f4ec15eed6f2b93412b6991ad26
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: hjk <hjk@qt.io>