qmlls: use radiobuttons for executable selections
Now that the user can choose the qmlls executable, change the qmlls
binary selection to use a radio button.
This simplifies the UI, as we don't have the weird combination of "use
latest qmlls" and "use custom qmlls" that can be valid at the same time,
and will be upgraded with a fourth radiobutton that allows to use (and
download) standalone versions of qmlls.

Task-number: QTCREATORBUG-32749
Task-number: QTCREATORBUG-33313
Change-Id: Icf6f81e0ae11ee9a723b2083ea3ef3caf866ada7
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: David Schulz <david.schulz@qt.io>