CmdBridge: Fix the watchdog killing the bridge unexpectedly
When either reading the command data or processing the command takes
longer than the timeout, the watchdog would kill the bridge because
pings cannot get through during that time. For that reason the timeout
was increased from a minute to an hour as a workaround, but it would be
better if the timer was reduced to a minute again and the watchdog
killed the bridge only if "the connection is lost", i.e. if no command
(even no ping) was handled for that time. So disable the watchdog timer
while the bridge is busy processing, and reset it afterwards.

This can for example happen when deploying larger files, where already
receiving the data (readPacket) can take longer than a minute.

Task-number: QTCREATORBUG-33317
Change-Id: I20135f6dd56b94aa16a9f1de7b64df1530ece6ea
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>