CmdBridge: Fix reading data packets
The packet structure is magicmarker + packetsize + payload.
We receive the data in chunks that can be broken up at any point within
that.
For the magicmarker and the payload we already handled that, but the
packetsize is also 4 bytes, and a chunk break can happen in there, so we
need to handle that too.

Fixes: QTCREATORBUG-33317
Change-Id: I4f77ca735d37b9679367ac2df559a65bbcda2e3e
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>