Improve order of items in context menu of file system view
The order of the "new", "rename", and "remove" items was a bit weird.
There was some kind of order by type, "file" versus "folder", but that
was broken by "rename" (which works on both) and confused by the
"Remove" items not always being visible.

Just group the two "new" items, followed by "rename" and the one
"remove" item corresponding to the current node.

Fixes: QTCREATORBUG-33270
Change-Id: I3212de118a10a29c480bec492ed175f076d6c1a9
Reviewed-by: Christian Stenger <christian.stenger@qt.io>
Reviewed-by: Robert Löhning <robert.loehning@qt.io>