CPlusPlus: Add parser exceptions for built-in type traits
In header files of compiler vendors, we find constructs such as this
one:
  template<typename T> requires __is_enum(T) struct E {};
This is rightfully rejected by our parser, as non-parenthesised function
calls are not allowed in a requires clause. However, as compilers
apparently have special handling for these built-in type traits, we
behave the same way as not to needlessly stumble over library headers.

Change-Id: Ib3199711dab706b8e27412f91f16f95748a1e3bb
Reviewed-by: Christian Stenger <christian.stenger@qt.io>