ProjectExplorer: Separate 'Source', 'Build' and 'Run' tools
Tools, also such on "devices" like docker operate in different
contexts, some work on the device directly when building on
this device (qmake, ...), some when things are run on this
device, and when (later) collecting e.g. (more) Qt versions it's
necessary to distinguish between versions meant to run on
a certain device or meant to be used for building for a
(possibly different) device.

I think it makes sense to separate these concepts also
visually (and perhaps later also filter out e.g. build stuff
from devices that will never used to build)

I've included also "Source" as processing sources does not
necessarily need a full build setup so it feels different,
too, and was easy to add. Might be over-engineered.

Change-Id: Ie6d86746b75274d82cef8f3378aadc8f9226f1a6
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>