Fix Coco plugin support for MinGW
The greatest problem was that the g++.exe wrapper was called by CMake
by its 8.3 file name (probably because of the "+" signs), so that it
could not find out as which compiler it should act. Now a --cs-profile
option is added for the C++ compiler under Windows.

Also some smaller corrections.

Fixes: QTCREATORBUG-33287
Change-Id: Icd91397e80980fa17c7bdac276133fe59e3ea799
Reviewed-by: Cristian Adam <cristian.adam@qt.io>