Add "Save Without Formatting"
The CppEditorDocument::saveImpl asks the indenter if formatOnSave is on,
and switches code paths for that. Indenter::formatOnSave() is a virtual
method that is implemented in the ClangFormat(Forwarding)Indenter and
checks the ClangFormat settings there.

The Beautifier plugin and the CMakeFormatter connect to
EditorManager::aboutToSave(IDocument*) and do it there.

Introduce an enum "SaveOption" with values for AutoSave,
DisableFormatOnSave, and None.

Use that instead of the "bool autoSave" for the various save
functions, and for the signals, so the corresponding code
can not only check if auto-saving is enabled in the corresponding
setting, but also if it was desired in the first place.

Add a "Save Without Formatting" action that triggers the save code paths
with SaveOption::DisableFormatOnSave.

Api: IDocument::save(FilePath, bool)
     -> IDocument::save(FilePath, SaveOption)
     IDocument::saveImpl(FilePath, bool)
     -> IDocument::saveImpl(FilePath, SaveOption)
     IDocument::aboutToSave(FilePath, bool)
     -> IDocument::aboutToSave(FilePath, SaveOption)
     IDocument::saved(FilePath, bool)
     -> IDocument::saved(FilePath, SaveOption)
Api: EditorManager::aboutToSave(IDocument *)
     -> EditorManager::aboutToSave(IDocument *, SaveOption)
     EditorManager::saved(IDocument *)
     -> EditorManager::saved(IDocument *, SaveOption)
Api: DocumentManager::saveDocument(IDocument *, FilePath, bool *)
     -> DocumentManager::saveDocument(IDocument *, FilePath,
        SaveOption, bool *)

Change-Id: I4b9d9a8b4855f87f1b70ee45edb15ba19cedc161
Reviewed-by: David Schulz <david.schulz@qt.io>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>