Examples: Filter Qt versions that do not have any documentation files
That filters out the boot2qt Qt version from the list of examples
on Windows & Linux. It sets QT_INSTALL_DOCS and QT_EXAMPLES_DOCS
to directories that exist and contain some other files. It does that
because these paths are used when installing examples on devices,
so we need to work around this in Qt Creator without actually doing
an actual scan for examples (which would be a performance impact).

Fixes: QTBUG-126753
Change-Id: Ib9952c6d8d20b7fcda7d0be42c1f1c95eac99689
Reviewed-by: Christian Stenger <christian.stenger@qt.io>