CMake: Disable SBOM generation for Creator repos unless opted in
To make it easier to incrementally add SBOM info to each Creator
plugin project, without breaking CI, we now skip SBOM generation for
repos that are not SBOM-ready, even if they are configured with
-DQT_GENERATE_SBOM=ON.

Consider a repo as SBOM-ready if it has a set(QTC_SBOM_READY ON)
assignment in its root CMakeLists.txt, before the
find_package(QtCreator) or include(QtCreatorAPI) call.
If the repo is not SBOM ready, the build system will force disable
SBOM generation even if QT_GENERATE_SBOM is ON.

Annotate the projects that can be built standalone as SBOM-ready
within this repo.

Change-Id: Icf5033aa03beb538f5f7600cba9682b06069fa2c
Reviewed-by: Eike Ziller <eike.ziller@qt.io>