CMake: Serialize install steps of wininterrupt and qtcreatorcdbext
A CI precheck shows a failure during the wininterrupt external project
win32-bld install step.

It fails while importing the 3rd party python dependency
'spdx_tools\spdx\parser\tagvalue\parsetab.py'
claiming a python syntax error.

It's unclear why this happens, given that the file is installed during
provisioning and is just executed during the Creator build.

One guess is that the spdx_tools scripts can't handle to be called in
parallel, which can happen when the external projects are installed by
ninja in parallel.

Add the USES_TERMINAL_INSTALL option to the external projects, to
serialize the install steps when using the Ninja generator.
Usually all CMake projects are installed serially by default, but
because we use ExternalProject_Add with custom installation steps,
we need to be explicit about it.

This is a speculative fix.

Change-Id: I2ba6686ac479d88f5c5281a6cc5e222a90195f11
Reviewed-by: Eike Ziller <eike.ziller@qt.io>