Editors/Tabs: Add Close All|Other Tabs
While the other close actions work on the editors/documents, add some
that work on the tabs of a single split.

"Close All Tabs" closes all tabs in the split, "Close Other Tabs" closes
all tabs except the selected one (either "current" or the one that the
context menu was triggered on).

Change-Id: Idca35f119b27cf5d728799e38a477078353c0d16
Reviewed-by: Christian Stenger <christian.stenger@qt.io>