Make it possibly to use rsync with non-local build devices
For example between Docker build device and remote Boot2Qt run device,
or between RemoteLinux and RemoteLinux devices.

`GenericDeployStep::transferTask` correctly chose rsync as the
preferred transfer method if the build device and the target device both
support it. But `createRemoteLinuxFileTransferInterface` fell back to
generic deployment anyway if the "source" isn't the local host.

The patch adds a "rsync on device" tool setting for devices. Changes the
check if the device supports rsync to take the tool setting into
account. Removes the disabling of rsync when creating the file transfer
interface for remote source files and fixes the file transfer function
itself to not make assumptions that it runs for the local host.

With this patch, with my Boot2Qt setup on macOS (Docker -> Boot2Qt),
after I disable "Run as outside user" and set the rsync tool path for
the docker device, open a terminal there and run

    apt update && apt install --yes rsync ssh && ssh-keyscan <target_ip>
        >> ~/.ssh/known_hosts

then Qt Creator uses rsync to deploy files between the Docker device and
the (remote) Boot2Qt device. Initial deployment of a large amount of
data still takes the same long time as with generic deployment (e.g. 5
minutes for 140 MB) due to copying it over the network, but with rsync
later deployments are much faster for files that didn't change.

Task-number: QTCREATORBUG-30831
Change-Id: Icea636659df4ea05558fee20f36398cb0f618175
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>