RemoteLinux: Ignore multi-line environment variables
... when prefixing the remote command with the environment.
Consider for instance this beauty, encountered in a Fedora environment:

BASH_FUNC_scl%%=() {  if [ "$1" = "load" -o "$1" = "unload" ]; then
 eval "module $@";
 else
 /usr/bin/scl "$@";
 fi
}

Trying to prepend that to the command line throws off the shell and thus
breaks command bridge execution.
Note that the conceptually correct thing to do would be not to set
redundant environment variables at all, i.e. only prepend the ones that
the user provided.

Change-Id: I1baf689d62f7b08b93936fdf32365591f760be12
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>