QmlJSEditor: Don't look for qmlls and qmlformat on remote installations
There presence is fairly expensive to evaluate, and this currently
happens synchronously for all Qt versions on Creator startup.

It's also not quite clear whether in case of a remote qmlls with
a higher version than the locally available ones the remote one
would be preferable, as there's also a significant run time overhead
when using.

In principle, this could be made working. For a proper solution we'd
need one qmlls per configured (build) device, not a single global one.

Since we are nearing QtC 18 feature freeze, but for pretend that
there's no qmlls for remote builds (yet).

Change-Id: I90ba09528688f65297a8df2d88a0e3306d556949
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>