McuSupport: Show only MCU kits when opening MCU demos and examples
MCUSupport is a CMakeProject, but it can only accept certain CMake
Kits.

It was not possible to intercept the project and change it's issue
in MCUPlugin before the kits filter logic has already passed.

The attempt was made using projectAdded which it self filters the
kits (unless aboutToAddProject signal is introduced)

We can't make MCUSupport a dependency to cmake project. so we only
rely on the projectPath for MCU's demos and example. it will be
more complete fix if we parse the file for keywords. but that would
be costly to do for each QtCreator kit in CMakeProject::projectIssues

Fixes: QTCREATORBUG-33352
Change-Id: I9b9edb154b6527f2855a0f54f67523f2b0f5c605
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>