Debugger: Fix Breakpoint file name set/retrieve
We want breakpoint filenames to reference the original file paths when
we set them in the debugger, so we have to map the paths accordingly.

When the debugger reports back a breakpoint we have to do the opposite
and try and find the local source of the file instead.

For that we store the build directory in the runParameters of the
debugger which will be a full device path to where the files are
actually built. It allows us to map our local file to where the file
will be on the build device (in case of a mount in a docker container
for instance).

Change-Id: Id6ac67b6b282bcaadd5a0533fbf0832feed156f8
Reviewed-by: hjk <hjk@qt.io>