FilePath: Fix resolution of symlinks with directory links
In the situation

    /a/file3.txt
    /x/linktest1.txt -> ../a/file3.txt
    /a/dirlink -> ../x

unfortunately symLinkTarget("/a/dirlink/linktest1.txt")
returns "/a/a/file3.txt", because readlink of that file returns
"../a/file3.txt" and that relative path is then resolved via the original
path, "/a/dirlink/../a/file3.txt".
It does not recognize that a directory link is involved that changes the
depth of the hierarchy.

That also breaks FilePath::resolveSymLinks that relies on symLinkTarget.

Fix it by checking the whole parent hierarchy of the path for symlinks
instead of just the file itself. If a link is found, that counts towards
the max 16 jumps that it does.

Adds some tests for resolveSymLinks.

Fixes: QTCREATORBUG-33335
Change-Id: Id55ae44cc4e8fabaf6ae37e87b0bc7d07e032d21
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>