RemoteLinux/rsync: Fix the ssh options for remote build device
The target device has an option for the host key checking, which
defaults to off. We need to pass ssh + these options to the rsync call
also when it is run on remote build devices.

Also add ssh as a tool that can be configured for devices. If it is
configured, that is used for rsync, otherwise it falls back to just
"ssh".

Task-number: QTCREATORBUG-30831
Change-Id: If762312c830a0f8adbe32d9265720ccccb1c9039
Reviewed-by: hjk <hjk@qt.io>