ProjectExplorer: add target for default kit for unconfigured workspace
Workspace projects are meant to be used with projects that do not have a
native support in Qt Creator, so by nature the common kit configurations
are usually not relevant for this project. Avoid confusing users by
having to choose any of the commonly Qt centric kits by always setting
up the default Kit when the project is opened for the first time.
Changing the kit later in the projects pane is still possible.

Change-Id: I72266af9eb4af3fe262497e3d4cccc499ee1c23c
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>