Devcontainer: Reload device when config is changed
When the configuration is changed, we ask the user if he would like
to restart the device with the new configuration.

To restart, we tear down the device and re-start it. Since tearing
it down removes the kits as well, the project will loose its targets.

We change it so that targets that are removed due to their kits being
removed are added as vanished targets. That way we can restore them
once the kit re-appears.

Change-Id: I0f5e498ee7b984a0b9e1479f5739dca33d8f4e63
Reviewed-by: Eike Ziller <eike.ziller@qt.io>