ProjectExplorer: Improve UI feedback of Kit enabling in Projects Mode
Using an additional Kit for the currently loaded Project requires one
click on the Kit item for enabling it and a second click for its
activation. The item visualization did not immediately reflect the
enabling, since the Kit icon remained gray and its "+" overlay did not
disappear.

With this change, the Kit item is immediately updated when enabled.

Fixes: QTCREATORBUG-33450
Change-Id: I7232c387c6815fe43362d3ff5b65d8d9f6577e19
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>