Fix the default deploy configuration
Amends e3a68d913343014349c1204e864f7135d1ccb5a2

which makes the generic deploy configuration available for all target
devices. Which as a side effect also made it the default for all target
devices (just because it is first in the list of available ones), which
is very wrong.

As an immediate fix prefer a any deploy configuration that is not
generically available for all devices as the default. Longer term it
might make sense to create a better prioritization, since at the moment
it is just sort of an accident that e.g. the Qt Application Manager
deploy configuration is not the default for all remote linux target
devices. (Well, the appmanager plugin depends on the remote linux
plugin, so the latter will always be initialized first, and the deploy
configuration appear first in the list, but...)

Change-Id: I930555cecbc501181f17fca5dec0589146838ddd
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>