Kits: Fix that selected value didn't match actual value
When changing e.g. the build device of a kit, the available items in for
example the toolchain and Qt version aspects are filtered.
The code in KitAspect::refresh tried to re-select the old value, but if
that was filtered out, it would just select the last ("None") item
available in the combo box, without updating the actual value in the
kit.

So if you added a kit (by default set up for desktop) and selected a
remote linux device, the toolchain and Qt version were visibly set to
"None", but the actual value for the kit (visible in the tool tip on the
kit) stayed on the desktop items. If then there is for example no Qt for
the device, it wasn't even possible to get rid of the Desktop Qt version
that was internally still set for the kit.

When a different item is selected in KitAspect::refresh, also adapt the
actual value in the kit.

Fixes: QTCREATORBUG-33456
Change-Id: I828f15e52e2c995aa0f7488831d40cc9d3c5f789
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>