Docker: Fix update of "reachable state" when mounts change
The kit is nowadays marked as "unsuited" for a project, if the build
device cannot "reach" the project directory.
When the Docker device's mounts change, this needs to be re-checked,
so send a DeviceManager::deviceUpdated signal to trigger the notification
chain.

Follows 4a8bb7ddbc7638de8af9734183615e9d89a2237d

Change-Id: Ic95a2b9d13ae57a35590bbb09e1e08c50149a6e1
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>