QmlDesigner: Reset puppet when first View3D is added
When the first View3D is added to the scene, the material is not
rendered properly in 3D view or material library until the puppet is
reset, so we now do that automatically.

Consolidated View3D and Model drops into single handler function in
Utils3D.

Now dropping View3D or Model into the code view also works correctly.

Fixes: QDS-15820
Change-Id: I75bf65e1e8b80d6cfb4d724468dabe0e805f66c2
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>