Overview: Remove duplicated example parsing code
Let the QtVersionManager do the example parsing with its regular means
instead. As a side effect this avoids that the Overview page tries to
access remote Qt versions and tries Qt versions in the same order as the
examples page. It also unifies parsing and opening of examples for the
future.

Since the overview page is already parsing for examples to decide if an
example should appear at all, we can also use that to avoid duplicating
the descriptions for examples.

Change-Id: I0599af637ca9718c54a8e2b729881d72da15fe21
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>