Tutorials: Work around differences in documentation URLs
The documentation page for the widgets tutorial moved to a different
URL. But the title stayed the same, and therefore the entry in the help
index for it stayed the same.

Add a little bit of a hack to interpret a "keyword" entry in the meta
data as an identifier for the page.

That corresponds to the "\keyword" command of qdoc, that adds an entry
in the index for the page. These are currently not exported to example-
manifest.xml by qdoc though.

Task-number: QTCREATORBUG-32772
Change-Id: Iedd054ccfe50d2b3ee8dd3ef1308fc14e0ff8d7d
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>