CMakePM: Fix incorrect resolution for inherited presets
Amends 7f1e16172a6a8a0a66c8505df80fcf156a26f894

As the CMake documentation states:

> If multiple inherits presets provide conflicting values for the same
field, the earlier preset in the inherits array will be preferred.

By traversing the list of presets in order, we achieve exactly this.

This change set adds a test for this bug, and one for the amended
commmit.

Fixes: QTCREATORBUG-32853
Change-Id: I4e1b2a0ce8cc5386efd8103095ce1b07d0330a40
Reviewed-by: Cristian Adam <cristian.adam@qt.io>