Examples: Fix the opened documents
The example manifests specify a list of files to open, and can specify a
"main file" that should be the one that is made active in case of
multiple files.

Since 018d49f417eb ICore::openFiles only actually opens the first file
it is given. The rest is opened as suspended documents. Adapt the code
accordingly.

Fixes: QTCREATORBUG-33266
Change-Id: I2db0fd73b7fc7eaffaa23b473bfd7031411fa09a
Reviewed-by: David Schulz <david.schulz@qt.io>