ProjectExplorer: Improve time to switch to Welcome mode again
Since the last rewrite we were reading session file contents
unconditionally for all sessions, running in some cases into
sessions that take 45ms to read their EditorData. With several
of such sessions this becomes user visible, in my case >300ms.

Fix this by only reading session data of sessions that are
expanded in the session list.

The is overall still a bit slower (~65ms) for me then it was
in QtC 16 (~30ms) but not noticeably so.

Change-Id: Id12af4157553eb5e5fb1c6a637276c910ba799e3
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>