FilePath: Fix test with Qt > 6.9.1
This is another fallout of the change in Qt that QTemporaryDir is no
longer a canonical file path on macOS, but /var/... (while the canonical
path is /private/var). But we need the assumption that the original file
paths for these tests are _not_ involving symlinks.

Amends 968135b74b660ac5849329c68276ded691c2f501

Change-Id: I0ee3e44b2db52182cce65d65cfebbf71b450886e
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>