CMake: Fix crash when switching build configurations
Amends e78975cfde18567692eba28202b977a392d37bac
that tried to fix a crash when disabling a kit while CMake configuration
is still running.

But the new connection guard fails to disconnect when the
CMakeBuildSettingsWidget itself dies, leading to even worse crashes.

The crash when disabling a kit happens because the parsingFinished
notification comes while the Target, BuildConfiguration, and
CMakeBuildSystem are already destructing.

Revert to the right connection guard, but let the code there be no-op if
we are already deconstructing everything.

Fixes: QTCREATORBUG-33517
Change-Id: I5f71d212a96f428513e4f5c77b42c0efd9dfe355
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Cristian Adam <cristian.adam@qt.io>