Android: Add the autocomplete functionality for the manifest editor
This adds the autocomplete functionality to the text editor widget in the
Android Manifest Editor. It has the most common attributes, elements
and permissions. The list is not exhaustive and will need to be maintained
manually. This does not check for placement validation, the suggestions
will show up regardless of whether it is inside "<>" and doesn't take
into account if the value is valid for that element etc.

Task-number: QTCREATORBUG-33250
Change-Id: I31476844d497773d33baa98d86d58fdd7dd6d9a0
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>