QmlDesigner: Remove secondary qml types from qmldir on unimport
If an imported content library user bundle item depends on secondary qml
types that are copied into the same bundle, those types are added to
bundle's qmldir file on import.
Now those types will be properly removed from the qmldir file on
unimport if the corresponding qml file is removed on unimport.

Fixes: QDS-15838
Change-Id: I20c8114a706f0261aebe41a46f824c72bd369124
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>