Fix importing iOS builds
It was always creating a temporary kit, instead of using the existing
iOS kits.

The iOS kits set the sysroot to
".../iPhoneOS.platform/Developer/SDKs/iPhoneOS.sdk" and
".../iPhoneSimulator.platform/Developer/SDKs/iPhoneSimulator.sdk"
respectively, because the compiler needs these flags passed, so this is
needed for the code model.

But the actual build sets CMAKE_OSX_SYSROOT=iphoneos / iphonesimulator
because that is how CMake does it.

The CMake build import just checked if the kit's sysroot matched the
CMAKE_SYSROOT, which failed.

Extract the CMAKE_OSX_SYSROOT from the build and do some rough checking
between the kit sysroot and that as well.

Change-Id: Ifafb0cb851c28f406bd2814a658bbcaff5844973
Reviewed-by: Cristian Adam <cristian.adam@qt.io>