Utils: avoid using QFontMetricsF::lineSpacing for block offset
QScriptLine that is used in the QTextLayout implementation only takes
positive leading into account when calculating the line height. If we
assume the QFontMetricsF::lineSpacing as the offset for unlayouted
blocks we get a lot unnecessary updates during the layout process. So
provide a static function in PlainTextDocumentLayout that provides a
lineSpacing that is closer to the lineSpacing in a plain text document.

Change-Id: I0ee114cf29f8a2fcbd5a2efe3fcea636a98c7739
Reviewed-by: Christian Stenger <christian.stenger@qt.io>