QmlDesigner: Use separate image container key for requests without id
Since property view can request a custom size of preview image
for a node and the image container key was only dependent on node id,
there was window of opportunity for shared image data meant for
material browser to be overwritten with different size image on puppet
side before it is read on qds side.

Custom requests now use separate key base to avoid this.

Fixes: QDS-14773
Change-Id: Ic5be292662c31dbc50435dc006d91488def8a017
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>