Avoid including multiple Tr classes
lupdate has issues finding the right translation context if there are
both Tr::tr calls that should be resolved via a `namespace <name> {`
scope, and another `using namespace <other>` with a namespace that has
another Tr class.

Many of the cases of this that we have were editor display names that
were defined in the Core plugin for reasons unknown, and which would
better be defined in the corresponding plugin anyway.

For the other cases do not use Tr::tr from a different plugin, but
instead introduce "msg<Something>" functions that do the Tr call plugin-
internally. That is the "correct" way to do it anyway, since shared
strings should be changeable centrally.

Task-number: QTBUG-140548
Change-Id: Idacf62b400f50789b5e5957b690e16aa1ebe5a56
Reviewed-by: Christian Stenger <christian.stenger@qt.io>