Editor: Fix updating horizontal x after delete
The QTextCursor::removeSelectedText indirectly emits
cursorPositionChanged before the horizontal x is updated and therefore
we do not recognize the changed x in TextEditorWidget since we connect
on cursorPositionChanged. So in order to get a correct x offset after
pressing delete we need to update the multi text cursor by ourself after
the deletion.

Fixes: QTCREATORBUG-33106
Change-Id: I64ff96902edac65073eb415c264caf683dec71d6
Reviewed-by: Christian Stenger <christian.stenger@qt.io>