Utils: Don't emit Aspect::checkedChanged() too often
If only the volatile value changed the value itself does not
have to have changed.

If the information that the volatileValue has changed were
needed, it should use a different (new) signal. Right now
both users of the signal are in an auto-apply setup, so
adding the signal is not (yet) needed.

Fixes: QTCREATORBUG-33449
Change-Id: I54085b790863dd5cbb8495540c861397e39e0f31
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: Orgad Shaneh <orgads@gmail.com>