Boot2Qt: Fix availability of the kit for projects
Amends e190a3e52ac79d762553605e8581b3e48c3fa913 which changed the logic
for the "unsuitable kits" to use IDevice::supportsBuildingProject,
which was added for DevContainer devices in
610f27b708797a6061079499defdf1aa446b874d but never for Docker devices.

The default in IDevice is to use handlesFile, which is wrong for Docker
devices, because that also supports projects that are not "on the
device" but mounted into it.

Add DockerDevice::supportsBuildingProject that gets the previous logic
back.

Fixes: QTCREATORBUG-33575
Change-Id: I385065537cc55cf51994b98acf2e51472fa7ff07
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: hjk <hjk@qt.io>