Editors/Tabs: Get wheel scrolling back
Amends 56fac32c7e191824f9588cf7da17448f02982ca8 in which we turned this
off, as it didn't work out of the box for us because it only called
setCurrentIndex, and seems to be a Qt-only feature.

But scrolling with the mouse wheel or touch pad does lessen the pain of
overflowing tab bars.

Since we use a subclass now anyway for the drag&drop support, add a
wheel event handler that calls the base implementation, and emits
tabBarClicked as well if any change happened.

Also turn on the scroll wheel handling for the editor tab bar on macOS.

Fixes: QTCREATORBUG-33568
Change-Id: I46e6af14363355d7f52d8ab0fd4dd9c1bf7faa5a
Reviewed-by: David Schulz <david.schulz@qt.io>