DeviceShell: Fix that responses could only be partially parsed
The setup of the device shell uses the process first to test for some
tools, and then proceeds with writing the shell script.

The responses to these individual commands were done just with
m_shellProcess->waitForReadyRead(), followed by
m_shellProcess->readAllRawStandardOutput/Error()

This has potential to fail, since even if there is output "ready" for
reading, that doesn't have to be a full line of output as it is expected
by the processing of the output read by readAllRawStandardOutput/Error

Make sure to try to read a full line from the process before parsing
that output.

Fixes: QTCREATORBUG-33495
Change-Id: I29d4c38d03125542ab75a2d480b72ea45c1e9312
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>