Core: Restore original width of ProgressView, right-aligned above popups
The now prevalent Popup notifications stretched the width of
ProgressViews and had them positioned left-aligned above, which looked
out of place.

The ProgressViews and the container of the Popup notifications were in
the same layout on the same level. This change introduces an additional
layout level which allows the stack of ProgressViews to have its
original maximum width and be right aligned above the other popups.

Task-number: QTCREATORBUG-33556
Change-Id: Ic829f730341a98b9d694dceb895915e6c329a711
Reviewed-by: Eike Ziller <eike.ziller@qt.io>