Learning: Design for "Recent Projects" list in Overview Welcome Page
The initial implementation re-used the existing recent project list. The
design in Figma has prevides a specific look for if, which this change
implements.

The specified "Show All" button, which would switch to the "Projects"
page is omitted though, and will be added later when there is API to
switch a welcome page.

Fixes: QTCREATORBUG-33551
Change-Id: I72665a6dede02f988aa060f0523a331c2c5d808e
Reviewed-by: Cristian Adam <cristian.adam@qt.io>