Utils: fix layout update ranges on document changes again
amends 37545e6d0060d644edf75995201cfb49f2126855

The assumption that from always points at the first character that
differs was wrong. The range can start with unmodified content. So in
total it is not save to base the end position for clear layouts on the
diff between added and removed characters, but we always need to clear
layout from 'from' to 'from + charsAdded'.

Change-Id: I5ad2678818b089f1ab722006bc1cb9e9ae9ec196
Reviewed-by: Christian Stenger <christian.stenger@qt.io>