qmllsclientsettings: make qmlls download more robust
Allow to re-download the qmlls client if a previous download failed
(that is, the directory with the version name exists but no qmlls
binary is in there).
Make evaluateGithubQmlls() filter out paths that don't contain a qmlls
binary, so that potentially broken qmlls downloads of newer versions are
ignored, and delete a potentially already existing download destination
folder before (re-)downloading qmlls.

Fixes: QTCREATORBUG-33592
Change-Id: Ie5365d2445fd13e29353621a02bf2909bfcb9246
Reviewed-by: David Schulz <david.schulz@qt.io>