QmlDesigner: Introduce NodeMetaInfo::inheritsAll
There is a demand to test if a sequence of types are inheriting a type.
For example to test if all types can be assigned to a list property.

Use would be:

auto typeIds = Utils::transform<SmallTypeIds<1024>>(modelNodes,
ModelNode::typeId);

bool inheritsAll = nodeMetaInfo.inheritsAll(typeIds);

We use type ids instead of NodeMetaInfos for less over head.

Change-Id: Ib84c58ba0e28aa0eec73b274696834b0c28f58eb
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>