Android: Add additional method of finding Android buildDirectory
After 6c2016677d0bf7e05a5b2cce56ebc05008686cda , buildDirectory() fails
to determine the build directory via the working directory (which
already was a hack), because workingDirectory of the buildTarget is not
set in multi-target projects.

This change appends an attempt to determine the build directory only in
case the other methods failed. Therefore, this should be a rather
"defensive" hack which does not break currently functioning
constellations.

Fixes: QTCREATORBUG-33569
Change-Id: I114fb837a66660d0c997685bd4f5b1f690857bd4
Reviewed-by: Cristian Adam <cristian.adam@qt.io>