QmlProfiler: Drop workaround for timeline height
The mentioned weird effects seem to have gone.

This prevents the user from scrolling too far where nothing is left to
display and keeps the screen nicely filled with information. If the
user tries to increase the size of the tool too much, then empty space
will simply fill the remaining gap at the bottom.

Change-Id: I81bceea04757a1ac9af90b7ee4fabc3347ded9a7
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>