QmlDesigner: Fix qml warnings on Texture property page
In TexturePane.qml, topSection element is in different id scope
than itemPane as it is inside a Component, so it is invalid id in
the scope of itemPane.
Changed refreshPreview to signal to work around this.
Similar pattern is used in MaterialPane.

Also fixed some runtime warnings in TextureSection.

Change-Id: I28a081f1c6e3a02880f0d5de490a6ebe70424c5e
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
(cherry picked from commit 9c99d23c8df10e5b53096cd86eccd81444e9b206)
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>