Lua: Fix possible nullptr access in Gui
This patch changes the Thing::ptr to be a QPointer, and adds a
onDestroyed field to the Lua bindings. It also adds a function that
checks any Thing derived object before accessing it to avoid crashes
when the underlying QObject has been deleted, e.g. because it was closed.

Since every Thing now has to build a QObject it is no longer useful
and is therefore merged into Object.

Api: LayoutBuilder::Thing base class has been merged into Object

Change-Id: Ie783242e738569cadede81a1c53ea5a8426a796a
Reviewed-by: hjk <hjk@qt.io>