Unify plugin meta data extensions to start with plugin ID
We have some meta data values that plugins can set to influence the
behavior of other plugins. Unify these to be located in a JSON object
with the reading plugin's ID as a key, so we don't end up with a
complete mess.

The old values are still read for compatibility.

Also document the meta data extensions that were added in the meantime.

Change-Id: Ice27d0738d330295eb55a023c61cc661a22e9626
Reviewed-by: André Hartmann <aha_1980@gmx.de>
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>