QmlDesigner: stop selection update timer when a new TextEditor is set up
When a new TextEditor is assigned, the selection update timer should not continue running.
This also applies when the new editor is nullptr, for example in:
TextEditorView::modelAboutToBeDetached()
or
where the connection to DesignDocument::designDocumentClosed triggers
m_widget->setTextEditor(nullptr).

This fixes a crash that occurs when the open project is switched very quickly.

Task-number: QDS-15654
Task-number: QDS-15983
Change-Id: Ia39ba0e36d28099fb62b2f9e2a569dabca391723
Reviewed-by: Marco Bubke <marco.bubke@qt.io>