CMakePM: Fix ninja detection on Docker / macOS
The issue with Docker was that `tool-
>filePath().deviceEnvironment().searchInPath("ninja")` would not have the
`scheme()` and `host()` that `tool->filePath()` has.

Now the code will search after ninja on a path that has the right
scheme() and host().

Amends 0291076e58171df206e64a29bf026e6bf3bc381b which added /opt/
homebrew/bin to the ninja search path.

This is not guarantening that CMake would find the ninja binary, due to
the fact that the GUI applications on macOS do not get the PATH
environment that the Terminal application would get.

This means that 0291076e58171df206e64a29bf026e6bf3bc381b didn't actually
fixed QTCREATORBUG-33560 (I was able to reproduce the issue with a clean
settings path after the "fix" was in).

Amends 96f11aae877698589ae4656e2cde008d8043f7a0 which removed the
canonicalization of CMake tools, which means that for macOS and `/opt/
homebrew/bin/cmake` it will stay like that, and not resolve to something
like `/opt/homebrew/Cellar/cmake/4.1.2/bin/cmake`.

The code in `CMakeGeneratorKitAspectFactory::addToBuildEnvironment` will
make sure that if `/opt/homebrew/bin/cmake` and `/opt/homebrew/bin/
ninja` are present, then `/opt/homebrew/bin` is added to the `PATH`
environment.

Fixes: QTCREATORBUG-33560
Fixes: QTCREATORBUG-33571
Change-Id: Ie34d7d0b22a62c9f454f13652eaaf48f70251477
Reviewed-by: Eike Ziller <eike.ziller@qt.io>